cadastro de paciente e endederecao fncionando

+++ clinica-medica/src/controller/FuncionarioDao.java
@@ -6,7 +6,6 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.time.LocalDate;
 
 import model.Endereco;
 import model.Funcionario;
@@ -28,7 +27,7 @@ public class FuncionarioDao implements IntefaceFuncionarioDao {
 			stm.setString(2, funcionario.getNome());
 			stm.setString(3, funcionario.getSexo());
 			stm.setString(4, funcionario.getTelefone());
-			//stm.setDate(5, funcionario.getDataNascimento());
+			// stm.setDate(5, funcionario.getDataNascimento());
 			retorno = stm.executeUpdate();
 
 		} catch (Exception e) {
@@ -44,7 +43,7 @@ public class FuncionarioDao implements IntefaceFuncionarioDao {
 	public boolean excluirFuncionario(Funcionario funcionario) {
 		con = Conexao.getInstacia();
 		Connection c = con.conectar();
-		
+
 		try {
 			String query = "DELETE FROM funcionario WHERE cpf = ?";
 			PreparedStatement stm = c.prepareStatement(query);
@@ -56,27 +55,26 @@ public class FuncionarioDao implements IntefaceFuncionarioDao {
 		} finally {
 			con.fecharConexao();
 		}
-		
+
 		return false;
 	}
 
 	@Override
 	public boolean alterarFuncionario(Funcionario funcionario) {
-		
+
 		con = Conexao.getInstacia();
 		Connection c = con.conectar();
 		PreparedStatement stm = null;
 		try {
-			stm = c.prepareStatement (
-					
-					"UPDATE seller SET nome = ?, sexo = ?, telefone = ?, data_nascimento = ? WHERE cpf = ?"
-					);
-			
+			stm = c.prepareStatement(
+
+					"UPDATE seller SET nome = ?, sexo = ?, telefone = ?, data_nascimento = ? WHERE cpf = ?");
+
 			stm.setString(1, funcionario.getNome());
 			stm.setString(2, funcionario.getSexo());
 			stm.setString(3, funcionario.getTelefone());
-			//stm.setDate(4, funcionario.getDataNascimento());
-				
+			// stm.setDate(4, funcionario.getDataNascimento());
+
 			stm.executeUpdate();
 			return true;
 		} catch (Exception e) {
@@ -84,7 +82,7 @@ public class FuncionarioDao implements IntefaceFuncionarioDao {
 		} finally {
 			con.fecharConexao();
 		}
-		
+
 		return false;
 	}
 
@@ -92,14 +90,14 @@ public class FuncionarioDao implements IntefaceFuncionarioDao {
 	public Funcionario consultarFuncionario(Funcionario funcionario) {
 		con = Conexao.getInstacia();
 		Connection c = con.conectar();
-		
+
 		try {
 			Statement stm = c.prepareStatement(null);
 			String query = "SELECT * FROM funcionario f INNER JOIN endereco e WHERE f.endereco_cep = e.cep";
 			ResultSet rs = stm.executeQuery(query);
-			
+
 			while (rs.next()) {
-				int cpf = rs.getInt("cpf");
+				Long cpf = rs.getLong("cpf");
 				String nome = rs.getString("nome");
 				String sexo = rs.getString("sexo");
 				String telefone = rs.getString("telefone");
@@ -108,30 +106,27 @@ public class FuncionarioDao implements IntefaceFuncionarioDao {
 				rs.getArray(endereco);
 				int numero = rs.getInt("numero");
 				String complemento = rs.getString("complemento");
-				
-				
+
 				Funcionario f = new Funcionario();
 				f.setComplemento(complemento);
 				f.setCpf(cpf);
-			//	f.setDataNascimento(dataNascimento);
+				// f.setDataNascimento(dataNascimento);
 				f.setNome(nome);
 				f.setNumero(numero);
 				f.setSexo(sexo);
 				f.setTelefone(telefone);
-				
+
 				Endereco e = new Endereco();
 				e.setBairro(complemento);
 				f.setEndereco(e);
-				
-				
-				
-			}			
-		} catch(SQLException e) {
+
+			}
+		} catch (SQLException e) {
 			e.printStackTrace();
 		} finally {
 			con.fecharConexao();
 		}
-		
+
 		return null;
 	}
